add finishJob API and some fixes

+++ src/main/java/com/fiseq/truckcompany/dto/TruckDto.java
@@ -23,5 +23,5 @@ public class TruckDto extends DtoBase{
     private HashMap<String,Object> truckModelAttributes;
     private String successMessage;
     private String errorMessage;
-    private Integer moneyLeftInAccount;
+    private double moneyLeftInAccount;
 }
